refactor: decouple APIs `keys()`, `values()`, `entry()` into a separate mod(file).
